Move activity utils to separate file for consistency
