+++ background/redux-slices/utils.ts
@@ -6,10 +6,6 @@ import {
   AsyncThunkPayloadCreator,
   createAsyncThunk,
 } from "@reduxjs/toolkit"
-import dayjs from "dayjs"
-import { convertToEth } from "../lib/utils"
-import { Asset } from "../assets"
-import { AnyEVMTransaction } from "../networks"
 
 // Below, we use `any` to deal with the fact that allAliases is a heterogeneous
 // collection of async thunk actions whose payload types have little in common
@@ -146,99 +142,4 @@ export function createBackgroundAsyncThunk<
     baseThunkActionCreator(action.payload)
 
   return webextActionCreator
-}
-
-function ethTransformer(value: string | number | bigint | null) {
-  if (value === null) {
-    return "(Unknown)"
-  }
-  return `${convertToEth(value)} ETH`
-}
-
-export type UIAdaptationMap<T> = {
-  [P in keyof T]?: {
-    readableName: string
-    transformer: (value: T[P]) => string
-    detailTransformer: (value: T[P]) => string
-  }
-}
-
-export type ActivityItem = AnyEVMTransaction & {
-  timestamp?: number
-  isSent?: boolean
-  blockHeight: number
-  infoRows: {
-    [name: string]: {
-      label: string
-      value: unknown
-      valueDetail: string
-    }
-  }
-  asset: {
-    metadata: Asset["metadata"]
-  }
-  fromTruncated: string
-  toTruncated: string
-}
-
-export function adaptForUI<T>(keysMap: UIAdaptationMap<T>, item: T) {
-  // The as below is dicey but reasonable in our usage.
-  return Object.keys(item).reduce((previousValue, key) => {
-    if (key in keysMap) {
-      const knownKey = key as keyof UIAdaptationMap<T> // guaranteed to be true by the `in` test
-      const keyAdjustment = keysMap[knownKey]
-
-      return keyAdjustment === undefined
-        ? previousValue
-        : {
-            ...previousValue,
-            [keyAdjustment.readableName]: keyAdjustment.transformer(
-              item[knownKey]
-            ),
-          }
-    }
-    return previousValue
-  }, {})
-}
-
-export const keysMap: UIAdaptationMap<ActivityItem> = {
-  blockHeight: {
-    readableName: "Block Height",
-    transformer: (item: number) => item.toString(),
-    detailTransformer: () => {
-      return ""
-    },
-  },
-  value: {
-    readableName: "Amount",
-    transformer: ethTransformer,
-    detailTransformer: ethTransformer,
-  },
-  gasUsed: {
-    readableName: "Gas",
-    transformer: ethTransformer,
-    detailTransformer: ethTransformer,
-  },
-  maxFeePerGas: {
-    readableName: "Max Fee/Gas",
-    transformer: ethTransformer,
-    detailTransformer: ethTransformer,
-  },
-  gasPrice: {
-    readableName: "Gas Price",
-    transformer: ethTransformer,
-    detailTransformer: ethTransformer,
-  },
-  timestamp: {
-    readableName: "Timestamp",
-    transformer: (item) => {
-      if (typeof item !== "undefined") {
-        return dayjs.unix(item).format("MM/DD/YYYY hh:mm a")
-      }
-      return "(Unknown)"
-    },
-    detailTransformer: () => {
-      return ""
-    },
-  },
 }
